requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3.56
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.3.1
matplotlib>=3.4.3
scikit-learn>=0.24.2
tqdm>=4.62.2 